+++ src/bornes/comandos/filters/mod.rs
@@ -1,5 +1,8 @@
 pub mod cargo_test;
+pub mod git_branch;
 pub mod git_diff;
 pub mod git_log;
+pub mod git_pull;
 pub mod git_status;
 pub mod pytest;
+pub mod table;